requirements.txt
kivy
pycups
pymupdf
img2pdf
fastapi
uvicorn
sqlalchemy
razorpay
pydantic-settings
jinja2
python-multipart
qrcode[pil]
aiofiles
python-dotenv
requests
